# This is a combination of 15 commits. # This is the 1st commit message:
[SPEC2DEF] Implement generating export patch information

Now spec2def uses the version annotations in spec files to generate a table, which will be used to dynamically patch the export table based on the current process' appcompat settings, when a dll is initialized, allowing to export the appropriate functions for each version.

This feature is still disabled in config.cmake with the variable DISABLE_EXPORT_VERSIONING.

# This is the commit message #2:

[NTDLL] Add mechanism to patch the export table at runtime

The loader checks for a ros-compat-descriptor, and if one is found, parses it's entries. Each entry corresponds to an export and provides a bitmask that specifies on what Windows version the export exists. If the bit for the process' appcompat version (default is still Windows 2003) is set, the export will be kept, otherwise the export will be moved into a second (private) export table. If a module has a ros-compat-descritor, it is still allowed to import these hidden exports.

# This is the commit message #3:

[ROSCOMPAT_APITEST] Add initial test for export rewriting

# This is the commit message #4:

[COMCTL32] Include roscompat descriptor

# This is the commit message #5:

[WINHTTP] Include roscompat descriptor,
so that winhttp can use nt6+ apis

# This is the commit message #6:

[REACTOS] Make sure xxx_vista is linked to before xxx

This is required, because with dynamically patched export tables we will export all vista functions from the main dll, and you can link to them at compile time, but they will not be resolved at runtime, unless the importer has a roscompat-descriptor. Linking the vista version first will resolve the imports from the vista dll.

# This is the commit message #7:

[NTDLL] Don't derive appcompat version from manifest

# This is the commit message #8:

[KERNEL32] Export more Vista APIs

Create kernel32_vista_static library and link both kernel32_vista and kernel32 to it.
Export some vista functions from kernel32.

# This is the commit message #9:

[NTDLL] Link rtl_vista and export some vista functions

# This is the commit message #10:

[APISETS] Improve/fix up some exports

# This is the commit message #11:

[MSVCRT] Export _vcwprintf

# This is the commit message #12:

[KERNEL32_APITEST] Hack the loader test

The test is mapping ntdll using MapViewOfFile, bypassing the loader, then enumerates it's exports, and then passes the names to GetProcAddress, which is hiding some of the original exports.
At this point we do not support this scenario.

# This is the commit message #13:

[SDK] Enable export patching

# This is the commit message #14:

[MSVCRT][CPPRT] Adjust Cxxhandler3

# This is the commit message #15:

oops