[NTOS:CM] Perform the writing/reading synchronously on log files
Whenever a read or write operation is done into a LOG file, it was done asynchronously. So what happens is that multiple operation requests touch the same LOG file at the same time and we end up corrupting the bitmap vector
of a registry hive when we're trying to find the set-bits from it, corrupting the stack (which is explicitly caught by _RTC_CheckStackVars).