[NTOS:MM] Increment balancer priority CORE-17624
Change KeSetEvent() increment to EVENT_INCREMENT in MmRebalanceMemoryConsumers() to ensure the balancer runs as soon as possible. MmRebalanceMemoryConsumers() is called by MiDecrementAvailablePages() when pages are critically low (MmAvailablePages < MmMinimumFreePages) meaning the balancer should be scheduled immediately, as there might not be enough pages for the balancer itself (ASSERT(MmAvailablePages >= 32);) later on.